allow mocking by returning interface types rather than final classes
